Fix up distributions for Block-Sparse Domains/Arrays
I'd failed to anticipate that if someone made a .distribution query on
a sparse block-distributed domain that we'd get an old _distribution()
type wrapping a BlockImpl class out, rather than the new Block type
we'd expect to get out.  This PR fixes that by adding the new
dsiGetDist() call to the SparseBlockDom class, similar to the BlockDom
class.

This then led to an internal error, which turned out to be due to a
code path in 'defaultSparseDist()' that was expecting all domain
values to have a .dist field that they could use directly and call
with _getDistribution().  Converting this to a call to the new
'.distribution' query that maintains backwards compatibility with
the old distribution types fixed that.

This led me to also notice that the deprecated '.dist' query would
also only do the old thing, where we'd want it to also use the new
'.distribution' logic, so I updated it to make that call instead (this
was due to bad/lazy merging on my behalf with the PR that deprecated
'.dist', though arguably that PR should've had the deprecated routine
call the new routine rather than duplicating its logic).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>