Reduce the number of copies of a 'Block' record made in common code patterns
In refactoring 'Block' to be a record, I inadvertently added an extra
copy of the 'Block' record when creating a new distribution for many
common code patterns.  This PR removes the extra copy by updating the
'chpl__buildDistValue() routine (which is intended to just be a
pass-through for the case of a record-based distribution) from:

  proc chpl__buildDistValue(x:record) {
    return x;
  }

to:

  proc chpl__buildDistValue(ref x:record) ref {
    return x;
  }

I'd naively assumed/hoped that with the default intents, copy elision
would simply have done this for me automatically, but no such luck.

I then had to do a similar update to the version of
chpl__buildDistDMapValue() that takes records for
backwards-compatibility.

This extra copy was only detected by one of our GPU tests because
the new copy of 'Block' resulted in an additional array assignment,
which resulted in an additional kernel launch.  But since the extra
copy of 'Block' results in an extra coforall+on to set up all of
the 'LocBlock' classes, I'm surprised that we didn't see any changes
in our correctness tests that do communication counts.

TODO:

- [ ] look into that last question of "Why didn't other things catch
  this?" a bit more

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>