Make chpl_library_finalize() not exit; Make chpl_library_init() error on 2x
This changes chpl_library_finalize() to not exit the Chapel program,
permitting the user to do other computations or cleanup activities
after finalizing the Chapel library, as requested in #22739.

It also adds an error if chpl_library_init() is called twice, in the
event that a user tries to re-initialize the Chapel library after
finalizing it (or simply tries to initialize it twice, which we
haven't guarded against in the past).

Resolves #22739.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>