Add an unstable warning to all log calls involving int and uint arguments
(except logBasePow2, which is already unstable)

log2 had explicit overloads to allow calls with `int` and `uint` arguments,
which would return ints.  This was not consistent with the behavior of the other
log functions, which did not have explicit overloads and so coerced such
arguments to reals of the nearest size.

W.r.t. 2.0, the plan is to mark all such calls as unstable.  This may mean that
the log2 overloads return reals in the future, or it may mean that log, ln,
log10, and log1p will change their return type when provided int or uint
arguments in the future.  We are short on time before the 2.0 release, though,
so will not make that decision yet.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>