Rename the conjg function to conj and adjust the already deprecated functions
Since the argument deprecation happened this release, we don't need a separate
deprecation for both the argument renaming and the function renaming, we can
just adjust the new version and the deprecation message, and the old version's
call to the new function.

And since the new version no longer only differs based on the argument, we no
longer need to mark the old version as "last resort".

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>