Move common code into DSIUtil module
I was originally thinking "This isn't _quite_ DSI related", but on
reflection, I think it's close enough to warrant being in the same
module (and maybe it's the module's name that should be updated).
This also has the benefit of not requiring us to change tests that
count dead modules and the like.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>