First try at GPU radix sort
This implements a radix sort that runs on GPUs with a correctness and
performance test for three different scan algorithms.
Currently we're doing the scatter operation (last step of radix sort)
on the CPU.
This was expected to change with atomics, but unfortunately the
algorithm does not allow for a stable behavior even with atomics, and
this is required for LSB radix sort. Therefore we need other ways to do
better.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>