Update other tests/module code; re-instate string argument to initializer
Some of these I missed, some I was dragging my feet on.

I hadn't realized that the lack of a string in the initializer would
break things, but it did.  From what I can tell, the string is unused,
so I'd probably remove it, but since I'm not focusing on stabilizing
this distribution, I'm not going to do that here.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>