Improve OFI landing zone sizing
OFI uses a "landing zone" into which incoming active messages are received.
The landing zone is split into two buffers used in a double-buffered
fashion -- messages are placed in one buffer until it is full as which point
libfabric switches to the other buffer. Once the active message handler has
finished processing the messages in a buffer it re-registers the now-empty
buffer with libfabric. Previously, the landing zone was 40MB. This size was
determined empirically on somewhat older hardware. Increase the default size
to 64MB and make it configurable via the CHPL_RT_COMM_OFI_LZ_SIZE environment
variable.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>