Add return values for GPU Atomics
Make the GPU atomic functions return the appropriate old values.
Currently these atomics are wrappers to nvidia or amd atomics, which do
return the old values, we didn't pass it along earlier.
Now we pass along the return value we get from the nvidia and AMD
atomic functions so we can use them in Chapel code.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>