Allow skipping 'kind' field for type constructors
This commit allows users to elide a value for the 'kind' field in a type
constructor. If the first non-named argument to a type constructor is
not of type 'iokind' or '_iokind', we will assume that the user intends
to write a type constructor expression without kind-related information.
For example, with this change a user can write ``fileWriter(false)``.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>