Create a gasnet-ex shim
Create a gasnet-ex shim that we can use to upgrade to the gasnet-ex API.
Today we use the gasnet-ex source, but the gasnet-1 API. We want to
start upgrading to the new API, but don't want to disrupt production
usage. To enable this, add a `CHPL_GASNET_VERSION` chplenv that can
either be `1` (the default) or `ex`. When `ex`, this will trigger using
the new `comm-gasnet-ex.c` source. Currently that's just a straight
clone, but we'll update over time. Note that `CHPL_GASNET_VERSION` is
part of the unique build path, so we can have gasnet-1 and gasnet-ex
builds in the same tree.

I considered doing this as a `CHPL_COMM=gasnet-ex`, but we have a lot of
things that check for `CHPL_COMM==gasnet` and it didn't feel worth the
time to track all those down and copy more of the runtime build
infrastructure.

Part of 14443

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>