Refactor Serializers and Deserializers to use return-helper
This commit refactors the DefaultSerializer and DefaultDeserializer
types and their 'start' methods to return a helper object. This helper
object will implement the methods necessary for a given type-kind (e.g.
Class, List, etc.), and will store any relevant state information inside
the helper object rather than the Serializer or Deserializer itself.

This approach allows for us to avoid creating a copy of the serializer
every time we invoke 'serialize', 'deserialize', or an initializer.

The API is approximately as follows, where '*' represents either the word
"read" or "write", for Deserializers and Serializers respectively. The first
row of "start" methods are implemented on Serializers or Deserializers
themselves, while the methods in other rows are implemented on the objects
returned by the "start" methods.

| Class      | Record      | Tuple      | Array      | List      | Map      |
|------------|-------------|------------|------------|-----------|----------|
| startClass | startRecord | startTuple | startArray | startList | startMap |
| *Field     | *Field      | *Element   | *Element   | *Element  | *Key     |
| startClass |             |            | startDim   |           | *Value   |
|            |             |            | endDim     | hasMore   | hasMore  |
| endClass   | endRecord   | endTuple   | endArray   | endList   | endMap   |

Note that the 'hasMore' methods are only available on Deserializer-helper
objects, so that users may query the helper object to dermine if they should
continue reading elements, keys, or values from the fileReader.

An example ``serialize`` implementation for a record with two integer fields
would now look like:

```chpl
proc MyRecord.serialize(writer, ref serializer) throws {
  var ser = serializer.startRecord(writer, "R", 2);
  ser.writeField("x", this.x);
  ser.writeField("y", this.y);
  ser.endRecord();
}
```

The return-helper for a class can now be passed upward to a parent serialize,
deserialize, or init method during (de)serialization. This is why the 'Class'
helper must implement another 'startClass' method itself. This pattern exists
to support formats that may care about how inheritance is represented. See the
test ``test/io/serializers/reader-init.chpl`` for an example usage.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>