Unstabilize range.first and range.last when applied to a range over bool or enum and that side of it is unbounded
From #20896 (comment)
For first and last, the name suggests "first/last in the sequence."
Given that the sequence is finite if idxType is finite (i.e. enum or bool),
it makes sense for first and last to be defined as if the range is bounded,
similarly to iteration.

To be on the safe side, we will mark this behavior as unstable.
Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>