Update IB testing scripts to run on Apollo system
For the most part, this is just renaming vars from CS to Apollo. The
main difference is that it's a pbs system, so instead of setting
partitions we have to set queues, number of cores, and launcher timeout.

We can't run ofi on a pbs system (we require a system launcher and PMI)
so we're removing that testing.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>