Prevent qthreads from stealing running tasks
Disable stealing of running qthreads tasks. We have runtime code that
relies on tasks not migrating to different pthreads (cache remote, parts
of ofi) so we disable features or fallback to a slower implementation
for (non-default) work-stealing qthreads schedulers where tasks can
migrate threads. Here as a compromise, prevent stealing of running
tasks. This should still allows us to load balance tasks that haven't
started when using a work-stealing scheduler.

Note that our default scheduler (nemesis) doesn't support work stealing,
so this really only impacts experimental schedulers like distrib. Even
for that we disabled work-stealing in 4505, but in the longer term we'd
like work-stealing and I think this is a good compromise.

The current motivator was that this allows us to remove an old patch
that forced tasks woken from a blocked sync var to return to the same
worker thread, since it was already coded to do that for unstealable
tasks. This allows us to stop carrying around an old patch and makes
upgrading qthread versions easier.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>