Fix compareExchangeWeak in atomic API tests
`compareExchangeWeak` is allowed to spuriously fail even when the
`atomic.val == expected`, this means we can't rely on expected not
getting updated from spurious failures and need to stop printing it and
reset it when it can fail spuriously. We were already looping to handle
spurious failures but didn't account for `expected` getting updated.

We haven't see this before because `compareExchangeWeak` doesn't fail
spuriously on x86, but I have hit this in some recent arm testing. A
sequence of instructions is used to implement `compareExchange*` on arm
and this sequence can fail for things like a context switch or if
another thread modifies the same cacheline. The atomics-api test was
failing for the promoted case where we have multiple tasks/threads doing
concurrent `compareExchangeWeak` calls on the same cacheline and causing
conflicts for each other. atomics-api would fail 1/10 times before, but
passes 10,000 trials on A64FX and ThunderX2 systems now.

Note that `compareExchange` (the strong version that can't spuriously
fail) is implemented with a retry loop under the covers and this is the
reason `compareExchangeWeak` exists since avoiding that second loop is
more efficient on platforms where the instruction sequence can fail.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>