Fix tests that assumed `c_char` was an alias for `int(8)`
The sign of `char` is implementation defined in C, but in practice is
signed on x86 and unsigned for arm (except for mac). We had a few tests
that we're baking in the signed type into .good files, which caused
failures on some recent arm testing. Fix tests here by:
 - Adjusting c_str_literal, cStringLiteral, and return-coerce-to-c-string to
   assert the type instead of print
 - Skiping localesForRegion since it's deprecated and going away soon
   anyways (type is in an error message.)
 - Removing unhandledType since I think unhandledTypeTuple still covers
   the spirit of the test.
 - Adjusting a C benchmark to use `int` for `getopt` instead of `char`

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>