Fix some filesystem portability issues in tests
Fix nonUTF8/basic.chpl when the test path is a symbolic link by using
`realPath` to construct an absolute path to a filename. Instead of
manually building a path to a known test location, you could just use
`absPath(filename)`, but that I wasn't sure if that would defeat the
purpose of the test (if `file.path` uses `absPath` under the covers.)

Also update nameLength/longName.precomp to call `getconf NAME_MAX` on
the current directory instead of on `/` since the root filesystem and
the current one can have different max name lengths.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>