Disable reading of maps with string keys/vals in default IO format
Our default IO format doesn't print quotes that we could use to
distinguish the beginning or end of a string, and so when reading the
default format we accidentally pick up separators like ':' or ','.
Instead, disable this ability for now so that users don't have to deal
with the non-functional cases.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>