Deprecate the name `cproj` and add a new function `riemProj` to replace it
Since we'd already deprecated the argument name this release, adjust the
deprecation message to instead talk about the function name and move the
implementation of the replacement so that it's sorted alphabetically after the
rename.

Also adjust the quick links at the beginning of the documentation so that they
refer to the new name instead of the deprecated one.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>