Deprecate the `carg` function fully, to be replaced by a new function `phase`
Since the argument name for the `carg` function had been deprecated this
release, we merely needed to update the deprecation message to include the name,
instead of having multiple deprecated versions around.

Alphabetizes the new version.  Removes the "last resort" pragma from the
deprecated version, as it was only needed when they shared the name function
name.

Updates all the documentation links to refer to the new name instead of the
deprecated version.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>