Adding new radix sort algorithm
Based on feedback from @mppf (thanks!), I wrote a version of the radix sort that uses different approach.
The first attempt wasn't parallel in the last (scatter) step, therefore this new approach was needed.
There is still work to be done to improve the performance of this new
approach.

Also added sort validation to perf testing and clean up test files

The diagnostics for existing tests change due to a change in array
initialization on scans.

More bugs might be excavated if different execopts are used. See code
notes for known possibilites

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>