Produce output required for XML test results
The JUnit XML test results are created by scraping the log, which requires
that tests produce the output expected by the scraper. The heterogeneous
processing unit test (cpuKind.py) wasn't doing that, so Jenkins didn't notice
that it was failing. Note that this commit does not fix the failure, I will
do that once it is confirmed that Jenkins now notices the failure.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>