Re-export ChapelIO symbols from IO module and mark the name ChapelIO unstable
If in the future we enable a way to opt out of including default-included
symbols, this will enable those symbols to be accessed again without using the
name ChapelIO.  This will also enable qualified access of those symbols today
using `IO` as the prefix, which is a stabilized name.

We don't like the name `ChapelIO`, but were having difficulty coming up with a
better one that was consistent with `AutoMath`.  This avoids resolving that
issue right away while not interfering with qualified access too much

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>