Change 'Block' to be an alias for 'blockDist' rather than 'BlockImpl'
I thought making 'Block' simply be the 'BlockImpl' class as it was
before would preserve historic behavior while generating deprecation
warnings, but that seems not to be the case, and I get simple cases
not compiling anymore.  Making it alias 'blockDist' instead seems
to work as expected, so I'm proceeding with that for now to see how
it works out.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>