Fix the config unstable test
The test was not throwing `--warn-unstable`, so when we stopped generating that
warning in all settings, it stopped turning up.  Additionally, by actually
throwing `--warn-unstable`, we discovered that the test needed an additional
warning because it was referencing the config in the body of the program as
well.  (Thanks Daniel for pointing that out!)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>