Fix a bug I introduced in #23270
Although #23270 fixed the behavior of the chpl_unwind.py script for
non-osx/linux systems, I accidentally broke it for osx/linux systems
that did not have an explicit setting made by removing the
fall-through logic that defaulted it to 'none', causing the script to
return 'None' when none of the other options were hit.  More proof
that I should never write Python scripts.

This fixes the error by moving the val or 'none' conditional that I
added last night out a level so that it's always used when none of the
other cases trigger.  Thanks to Jade for pointing out the error in my
control flow.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>