Test modifications to avoid failures
These tests explicitly throw both `--memLeaks` and `--warn-unstable` normally,
so when they are run with a general `--memLeaks` paratest, we end up getting
two warnings instead of the single one they were expecting.  This is generally
good, as it means we are triggering the warning for every flag, but it means
we'd be back to need a separate .good file for the general memleaks run.  So
instead skip in that case (as we are already covering their behavior in normal
testing, because the flag is explicitly always thrown)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>