Adjust gcd implementation and add overloads to return different sized ints/uints
Damian requested a more generic and tighter version of gcd for 32 bit and 16 bit
integers.  While smaller sized ints could be sent to the original implementation
just fine because of coercion, it would always return the default-sized int.

Adjust the implementation for the default version to use his suggested code, and
add explicit overloads for `int(32)`, `int(16)`, `int(8)`, `uint(64)`,
`uint(32)`, `uint(16)`, and `uint(8)`.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>