Generalize and tighten up read/writeBinary() for arrays
This changes how read/writeBinary() interacts with arrays by:

* extending them to work with multidimensional arrays
* restricting them to work only with contiguous, local arrays
  - and slices thereof
* adding errors for cases we can't handle to avoid accidentally
  getting a promoted call to a scalar routine

The first bullet was proposed in #23081 and generally accepted by the
IO team since it seemed increasingly arbitrary to support 1D arrays
only.

The second bullet, I believe, reflects what was intended for these
routines on arrays when Michael and I were originally discussing them,
which was to limit them to fairly primitive array types that could be
blasted out/in with a single write/read.  Though the routines accepted
Block-distributed arrays (say) before this PR, the read() didn't
actually work, and the approach relied on reading/writing element by
element.

Currently, if the array being read/written is not DR/a DR view,
we fall into the error overload and complain about the array not
currently being supported.

If the array is not contiguous, we throw an illegal argument error.

The existing logic also didn't handle the read/write using a single
data blast for endiannesses other than native, so I also added a
compiler error for those cases until such time as we can do it in a
single read/write.

TODO:
- [ ] are we OK with restricting to native endianness for the time being?
- [ ] wording/phrasing of illegal argument error?
- [ ] how to document helper routine used in where clause?
- [ ] how to document what we do or do not support?
- [ ] more testing, particularly for slice cases (e.g., column or row
      of local 2D array?  local slice of block-distributed array?)

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>