Remove uses of useNewOpenReaderRegionBounds from file.readerHelper
Now that the flag is meaningless, we do not want to check against it, we instead
want to always provide the behavior that setting it to `true` provided.  Update
comments that reference it.

I'm hoping to deprecate the fileReader flag as well, but might run out of time,
so try to ensure we have a working product in the interim.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>