Restore bool->bool cast
This restores the bool->bool cast that I removed when removing
'bool(w)', thinking it was pointless.  It turns out that it mostly is,
except that it was the only thing keeping an Arkouda
'(array-of-bool):bool' cast working.

Restore it for now, though I think the more proper thing to do is to
add an 'array-of-T:T' cast to handle such patterns in general.  That
is, it should not be required to create a self cast in order to do a
no-op cast on an array, I think.  Will pursue that question and effort
separately.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>