Remove -suseNewFileReaderRegionBounds=true from these tests
For the test-rmatalt directory, that was the only contents of the COMPOPTS file,
so delete it.  For the other tests, there were other options provided, so merely
remove the no-longer-necessary flag.

This will not address all the tests that use this flag, snapshotting so that I
can go back to fixing the memleaks failure

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>