Remove deprecated file.lines region argument behavior and deprecate flag
This behavior has been deprecated for a few releases now, so can be removed.
Deprecate the flag that enabled switching between the behaviors and remove its
uses.  Also remove where clause on the remaining version of the function, since
it was solely added to avoid warning when no explicit high bounds was provided

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>