Remove deprecated high bounds behavior from file.writer and deprecate flag
This behavior has been deprecated for several releases now, so it can be
removed.  Deprecate the flag that controlled it, and remove uses of that flag.
Remove versions that were supporting the old behavior and fully remove where
clauses from the remaining versions, as the where clauses were added to avoid
generating warnings when the high bound was not set.  Also removes the
fromOpenUrlWriter argument from writerHelper, as it was used to given
openUrlWriter consistent behavior.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>