Remove the deprecated version of j1 and move its implementation into Math
We deprecated including j1 by default in all user programs a few releases ago,
so it is safe to remove that version of the function.  Its implementation was
still living in the AutoMath module so that it did not have to use the Math
module for the function to continue to work and to avoid code duplication/drift.
Now that the deprecated version is getting removed, we can put the
implementation back in the version that will be kept

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>