Update test to avoid leaks
PrivateDist leaks (on main and in this branch), so this puts all of
the interesting test code into a conditional to prevent it from
running while still checking that it gives the expected deprecation
warnings.  Meanwhile, the code outside the conditional checks that
the type exists and works in a very limited (but leak-free!) fashion.

The reason we've been getting away with this is that what few
uses of PrivateDist there are use the PrivateSpace singleton
rather than creating their own instances of PrivateDist.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>