Convert DimensionalDist2D to a record
This was more challenging than the others because
* it's highly generic
* its code is structured very differently

Most tests seem to be working, though I had to dodge some things
that didn't:

  new dmap(new DimensionalDist2D(...))

didn't seem to work as in other cases...  Removing the 'new dmap()'
did work though.

Doing an init copy of the record as in t1.chpl and t9.chpl didn't
work.  I used a ref instead.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>