Fix links from per-dimension distributions back to DimensionalDist2D
I couldn't figure out why these weren't working last night but Jade
solved it this morning: Because chpldoc ignores the type of symbol
field (e.g., ':class:'), we'd been linking to the module rather than
the type in question, and I couldn't link to the type without
providing a module path.

After implementing that fix and looking at the resulting docs, I
decided that linking to the module itself is the better thing to do
since (a) the type's the first/only thing in the module and (b)
linking to the type causes things like the instability warning to
scroll off the screen.  So here, I kept the link to the module, fixed
the white lie that we were relying on before last night's change, and
renamed the link to still name the distribution.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>