Fix bug with hasFirst, hasLast, and isEmpty
These functions on ranges had behaviour inconsistant with range.first
and range.last for unbounded ranges over enums/bools
We marked range.first and range.last as unstable over ranges over
enums/bools and that side of the range is unbounded.
This is becuase our current working theory is to treat ranges over
enums/bools as always representing a bounded range since the types are
finite. based on this hasFirst, hasLast, and isEmpty should also be
unstable in those cases and follow the correct behavior

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>