Keep submitted shootouts passing testing
With so many submitted shootouts relying on removed deprecated
features and not passing testing, we lose the ability to make sure we
haven't lost performance or features.  This PR keeps the tests limping
along, for now with minimal source changes from what was submitted, by
adding a 'Compat' module that implements all missing features in terms
of other features and 'use'-ing it at the end of the file to keep the
diffs restricted to one line.  This should permit them to continue
being tested until we have the chance to rewrite them, ideally in a
2.0-compliant manner.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>