Split Compat modules into three — normal, GMP, and RE2
In #23428, I failed to recognize the obvious — that in introducing my
'Compat' module, I was infecting all tests that used it with potential
dependencies (GMP, RE2) that their skipifs wouldn't skip past.  Here,
I'm splitting the Compat module into three so that each test can only
bring in those third-party dependencies that it was already expecting
to use.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>