Avoid warnings about implicitly converting ints to uints in the meteor benchmark
Avoids the warnings by adding a cast to each place pointed out by the warning.
An alternative to consider is looking into making more of the variables used in
this benchmark be `uint` to begin with.  Theoretically, this shouldn't pose a
problem as I think all uses are purely positive.  But I didn't take the time to
do that investigation.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>