Convert gasnet-ex shim barrier over to the GEX API
Convert barrier over to the GEX API. GASNet-EX barriers are anonymous
and use the new collective interface and standard GEX events to check
for completion. Note that unlike `gasnet_barrier_try`, `gex_Event_Test`
does not poll, so here we need to manually poll if we're barrier'ing
early in init before the progress thread has been started or late in
deinit after the progress thread has stopped.

Since barriers are now anonymous, it's not possible to receive a barrier
mismatch error, which I think was the only interesting error case
before. This enables us to remove the IMO confusing error checking that
was done before (confusing because the macro looked like we were making
another barrier call, even though we were just using the string for
error messages.)

Part of 14443

Co-authored-by: Dan Bonachea <dobonachea@lbl.gov>
Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>