Retire the FFT benchmark from release/examples, moving it into the studies dir
Removes it from the Makefile and README in that directory.  Adjusts its compopts
and perfcompopts to continue referring to the HPCCProblemSize.chpl helper file
in the release/examples directory without copying the file.  Adjusts the
GRAPHFILES and ML-GRAPHFILES to refer to the new .graph file locations instead
of the old ones.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>