Convert gasnet-ex shim to use real GEX broadcast for seginfo broadcast
Convert the naive AM-based broadcast of the seginfo table done for
segment everything into a real collective broadcast. This is cleaner and
more efficient (and a TODO from 7395837). As part of this, don't
broadcast zero'd values, just broadcast locale-0 values and
independently zero non locale-0 entries.

While technically more efficient, this is a single small broadcast so
doesn't have a noticeable impact on startup. With 100 nodes and 800
processes (`-nl 100x8` with support for process per numa domain) using
ibv with segment everything I see the barrier+broadcast+barrier is sub
millisecond with either approach. This single use is pretty noisy, but
adding an artificial 1,000 trial loop I see timings per iteration go
from ~800 microseconds to ~75 microseconds, so there is a real impact of
using a collective, but as a single use the old method wasn't horribly
slow either. Still, this is a nice code cleanup, so worth doing.

Part of 14443

Co-authored-by: Dan Bonachea <dobonachea@lbl.gov>
Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>