Minor primer index adjustments
It turns out when a primer contains top-level section headers set off
by `======` they appear on the primers table of contents page as well
(?!), so this steps all headers in the loops primer down a level to
fix this.  I apparently hadn't looked at the index in awhile.

While here, I moved the distributions and forall loops primers up in
their section because I think they're more prominent now than some of
the others, particularly due to the improved discussion of distributed
arrays and domains, and the growing importance of task intents.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>