Retire LCALS from release/examples and move it to the studies directory
We're no longer confident that the implementation of this benchmark reflects
good Chapel style and practices.  Until such time as we can go through and
review it (or make a new version), retire it from our recommended examples.

Removes it from the README in release/examples/benchmarks.   Updates GRAPHFILES
so that it's no longer in the Release examples or Performance Tracking suite and
so the remaining references use the right location.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>