Most of the code for the October 16th tutorial.
Everthing compiles and runs.  I have a list of things that have unstable
things in it, but some of it is things like module Random and GPU code,
which will still want to use in the tutorial.

Will be asking for tutorial ad hoc subteam members to help put together
needed .good files so can get rid of NOTEST file.

Still missing a couple of things from decks 03 and 04.

Jeremiah should merge in the heat 1d and 2d code here.
---
Signed-off-by: Michelle Strout <michelle.strout@hpe.com>