Change the regionArgReader test to be more reliable
Uses a local http server we set up ourselves locally instead of one hosted on
the web.

As a result of making this change, I had to change the starting location to look
at, since I wasn't sure how to make the file we're getting seekable (so we can't
start at anywhere other than the first point in the file).

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>