Fix method scopes for sibling/field calls
This commit allows for methods and fields to be used in the signature of
methods defined on the same aggregate type. The
``Resolver::createForInitialSignature`` method is updated to resolve a
``this`` formal in the case of methods, and used the resolved type as
the Resolver's ``inCompositeType`` value.

Add ``Resolver::setCompositeType`` as a means of setting the Resolver's
composite type once the ``this`` formal has been instantiated.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>